LPS-180422 Move to the API module as it will be used from Objects
