+++ modules/apps/portal-vulcan/portal-vulcan-api/bnd.bnd
@@ -18,6 +18,7 @@ Export-Package:\
 	com.liferay.portal.vulcan.graphql.util,\
 	com.liferay.portal.vulcan.graphql.validation,\
 	com.liferay.portal.vulcan.jaxrs.exception.mapper,\
+	com.liferay.portal.vulcan.jaxrs.extension,\
 	com.liferay.portal.vulcan.list.type,\
 	com.liferay.portal.vulcan.multipart,\
 	com.liferay.portal.vulcan.openapi,\
